Refactoring. Implementing dependency injection and better typehinting. Use of pydantic for data validation. In progress. TODO, implement side models using pytorch and keras for performance comparison.

+++ src/model/model/sequential.py
@@ -68,7 +68,7 @@ class Sequential(Model):
                          f"Output Shape: {layer.output_shape}")
 
     # <-- Compile Model -->
-    def compile(self, loss: str, optimizer: str, learning_rate: int = 0.001):
+    def compile(self, loss: str, optimizer: str, learning_rate: float = 0.001):
         """
         Configure the model for training.
